requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
bandit==1.7.8
black==24.3.0
cachetools==5.3.3
cerberus==1.3.5
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.4.4
defusedxml==0.7.1
distlib==0.3.8
django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djantic==0.7.0
execnet==2.1.1
filelock==3.13.3
flake8==7.0.0
gprof2dot==2022.7.29
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
hypercorn==0.16.0
hyperframe==6.0.1
idna==3.6
iniconfig==2.0.0
jinja2==3.1.3
kolo==2.23.0
lxml==5.2.1
markdown-it-py==3.0.0
markupsafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
msgpack==1.0.8
mypy-extensions==1.0.0
nose2==0.14.1
outcome==1.3.0.post0
packaging==24.0
pathspec==0.12.1
pbr==6.0.0
pip==24.0
platformdirs==4.2.0
pluggy==1.4.0
priority==2.0.0
py-cpuinfo==9.0.0
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.2.0
pygments==2.17.2
pyproject-api==1.6.1
pysocks==1.7.1
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-benchmark==4.0.0
pytest-cov==5.0.0
pytest-django==4.8.0
pytest-html==4.1.1
pytest-instafail==0.5.0
pytest-metadata==3.1.1
pytest-mock==3.14.0
pytest-profiling==1.7.0
pytest-sugar==1.0.0
pytest-timeout==2.3.1
pytest-xdist==3.5.0
pyyaml==6.0.1
requests==2.31.0
responses==0.25.0
rich==13.7.1
ruff==0.3.5
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
stevedore==5.2.0
termcolor==2.4.0
toolz==0.12.1
tox==4.14.2
trio==0.25.0
trio-websocket==0.11.1
typing-extensions==4.10.0
tzdata==2024.1
ulid-py==1.1.0
urllib3==2.2.1
uv==0.1.28
virtualenv==20.25.1
wsproto==1.2.0
